Massive cleanup of the project overall
