+++ .gitignore
@@ -15,6 +15,9 @@ dist
 dist-ssr
 build
 
+# Generated protobuf output
+src/api/generated/
+
 # Env files
 .env
 .env.*
