+++ README.md
@@ -22,6 +22,18 @@ Go to https://cloud.maptiler.com/maps/
 <br>
 To generate an API key.
 
+## Protobuf
+
+For the project to work you need to have a generated api_pb.ts file.
+<br>
+You can generate this by running this command:
+
+```bash
+npm run proto:generate
+```
+
+**MAKE SURE THAT THE PROTOBUF FILE MATCHES THE C# PROJECT**
+
 ## Running the project
 
 Command:
