+++ src/api/client.ts
@@ -1,5 +1,7 @@
 const API_BASE_URL = 'http://localhost:5000';
 
+
+// Generic GET request function that returns a typed response
 export async function apiGet<T>(path: string): Promise<T> {
   const response = await fetch(`${API_BASE_URL}${path}`);
 
@@ -10,7 +12,9 @@ export async function apiGet<T>(path: string): Promise<T> {
   return response.json() as Promise<T>;
 }
 
-export async function apiPost<T>(path: string, body: any): Promise<T> {
+
+// Generic POST request function that sends JSON and returns a typed response
+export async function apiPost<T>(path: string, body: unknown): Promise<T> {
   const response = await fetch(`${API_BASE_URL}${path}`, {
     method: 'POST',
     headers: {
@@ -24,4 +28,26 @@ export async function apiPost<T>(path: string, body: any): Promise<T> {
   }
 
   return response.json() as Promise<T>;
+}
+
+
+// POST request function that expects a binary response (e.g., for protobuf data)
+export async function apiPostBinary(
+  path: string,
+  body: unknown
+): Promise<Uint8Array> {
+  const response = await fetch(`${API_BASE_URL}${path}`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(body),
+  });
+
+  if (!response.ok) {
+    throw new Error(`Request failed: ${response.status} ${response.statusText}`);
+  }
+
+  const arrayBuffer = await response.arrayBuffer();
+  return new Uint8Array(arrayBuffer);
 }
